+++ packages/pluggableWidgets/gallery-web/src/components/GalleryContent.tsx
