+++ packages/pluggableWidgets/gallery-web/src/components/GalleryFooter.tsx
