+++ packages/pluggableWidgets/gallery-web/src/components/GalleryHeader.tsx
