+++ packages/pluggableWidgets/gallery-web/src/components/GalleryTopBar.tsx
